Carefully synchronize the threads in order to run a minor collection

+++ c7/stm/pages.c
@@ -8,7 +8,7 @@ static void pages_initialize_shared(uintptr_t pagenum, uintptr_t count)
     /* call remap_file_pages() to make all pages in the
        range(pagenum, pagenum+count) refer to the same
        physical range of pages from segment 0 */
-    long i;
+    uintptr_t i;
     for (i = 1; i < NB_SEGMENTS; i++) {
         char *segment_base = get_segment_base(i);
         int res = remap_file_pages(segment_base + pagenum * 4096UL,
